Added invocation results, confirmed that custom sounds work!
Added invocation results for all of the features that I have been working on that didn't already have results. This will hopefully make it more clear when stuff works/doesn't work. Right now it still seems there is a requirement to reload the apk from backup every time instead of repatching an already patched apk. This may need to get addressed, or it might be a zip problem.